+++ commands/ban-management.js
@@ -4,6 +4,8 @@ const xmpp = require('node-xmpp-client');
 const Assets = require('../utils/Assets');
 const Websocket = require('../utils/Websocket');
 const Say = require('../utils/Say');
+const Log = require('../utils/Log');
+const Client = require('../utils/Client');
 const Settings = require('../utils/Settings');
 const Templater = require('../utils/Templater');
 
@@ -88,6 +90,13 @@ class BanManagement {
      * @return {void}
      */
     static banUser( username, chat ) {
+        // Don't try to ban the streamer or the bot
+        let user = Client.getUser( username );
+        if ( user.isStreamer() || user.isBot() ) {
+            Log.log('Attempt to ban streamer or bot.' );
+            return false;
+        }
+
         let affiliationStanza = BanManagement.getUserAffiliationStanza( chat.credentials, username, 'outcast' );
         chat.client.send( affiliationStanza );