+++ commands/say.js
@@ -15,12 +15,12 @@ module.exports = [{
     action: function( chat, stanza ) {
         // Parse the message from the command,
         // limit !say message to 80 chars
-        var message = regex.exec( stanza.message )[2];
+        let message = regex.exec( stanza.message )[2];
 		message = message.substr( 0, 80 );
 
         // Allow users to override the voice
-        var voice = defaultVoice;
-        var match = /^\-voice\s(\w+)\s(.+)/.exec( message );
+        let voice = defaultVoice;
+        let match = /^\-voice\s(\w+)\s(.+)/.exec( message );
         if ( match ) {
             voice = match[1];
             message = match[2];