+++ changelog.md
@@ -2,6 +2,14 @@ Starting to keep track of changes as of October 28, 2015. I will try my best to 
 
 You can also browse the commit history to track my commits: [https://github.com/owenconti/livecodingtv-bot/commits/master](https://github.com/owenconti/livecodingtv-bot/commits/master)
 
+## Nov 7 2015
+* Added a setting for the `unavailable` command to enable/disable the command.
+* Added a `setsubject` command, to set the room's subject via XMPP.
+* Added the ability to pass a settings object to the `client.js` code of plugins.
+* Updated `help` command to output a link to Pastebin when called.
+* When the bot starts, it will generate and store the documentation for help in a Paste on pastebin.com.
+* Greetings are now run through a templater with `username` and `status` variables being passed-in.
+
 ## Nov 5 2015
 * Say command messages limit increased to 80 characters.
 * Added a Say message to the ban command.